graduate-work/requirements.txt
alchemy-mock==0.4.3
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dataclasses==0.8
Flask==1.1.2
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.9.0
marshmallow-sqlalchemy==0.24.0
mccabe==0.6.1
pyaml==20.4.0
pylint==2.6.0
python-dotenv==0.15.0
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
yapf==0.30.0
